+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 * added `MockTrait`, `ComponentTestCase` and `HelperTestCase` classes for test
     suite;
 * added `getParserDescription()`, `getParserOptions()` and
-    `getParserSubcommands()` methods for the `ConsoleIntegrationTestCase` class.
+    `getParserSubcommands()` methods for the `ConsoleIntegrationTestCase` class;
+* added `TestCaseTrait::assertIsMock()` assert method.
 
 ### 2.17.4
 * fixed bug in the integration with CKeditor when uploading images (see